+++ .gitignore
@@ -1,0 +1,3 @@
+Tester_cc.o
+Tester_c.o
+libtester.so
