+++ Answer.hpp
@@ -1,0 +1,7 @@
+#if !defined(ANSWER_HPP)
+#define ANSWER_HPP
+#include <stdint.h>
+
+typedef uint8_t Answer;
+
+#endif // ANSWER_HPP
