+++ Question.hpp
@@ -1,0 +1,15 @@
+#if !defined(QUESTION_HPP)
+#define QUESTION_HPP
+#include <stdint.h>
+#include <string>
+
+struct Question
+{
+    string first;
+    string second;
+    string third;
+    uint8_t number;
+};
+
+#endif // QUESTION_HPP
+
