+++ Tester.cpp
@@ -1,5 +1,7 @@
 #include "Tester.hpp"
 
+static const string Tester::INSTRUCTION = "Some instruction";
+
 Tester::Tester(){
 
 }
@@ -26,19 +28,3 @@ void Tester::SaveCurrentAttempt(){
 
 }
 
-Attempt::Attempt(){
-
-}
-
-Attempt::~Attempt(){
-
-}
-
-Question* Attempt::NextQuestion(){
-    Question* question = new Question();
-    question->first = "first";
-    question->second = "second";
-    question->third = "third";
-    question->number = 30;
-    return question;
-}
