+++ Tester.hpp
@@ -5,16 +5,16 @@
 #include <stdint.h>
 #include <string>
 
+#include "Attempt.hpp"
 #include "Storage.hpp"
 
 using namespace std;
-typedef uint8_t Answer;
 
 class Tester
 {
 private:
     /* data */
-    static const string instruction = "Some instruction"; 
+    static const string INSTRUCTION; 
     std::auto_ptr<Storage> storage;
     Attempt* currentAttempt = nullptr; 
 public:
@@ -24,38 +24,6 @@ public:
     Attempt* GetAttempt();
     static const string& GetInstruction();
     void SaveCurrentAttempt();
-};
-
-class Attempt
-{
-private:
-    /* data */
-    std::auto_ptr<Person> person;
-    uint8_t currentQuestion;
-    vector<Answer> answers;
-public:
-    Attempt();
-    ~Attempt();
-    Question* NextQuestion(); // TODO: alexEP
-    void GiveAnswer(Answer answer); // TODO: alexEP 
-};
-
-
-struct Question
-{
-    string first;
-    string second;
-    string third;
-    uint8_t number;
-};
-
-struct Person
-{
-    string lastName;
-    string name;
-    string email;
-    int64_t birthDate;
-    int64_t testDate;
 };
 
 #endif
